Module eosio_sys::ctypes

Structs

CStr
Representation of a borrowed C string.

CString
A type representing an owned, C-compatible, nul-terminated string with no nul bytes in the middle.

FromBytesWithNulError
An error indicating that a nul byte was not in the expected position.

IntoStringError
An error indicating invalid UTF-8 when converting a CString into a String.

NulError
An error indicating that an interior nul byte was found.

OsStr
Borrowed reference to an OS string (see OsString).

OsString
A type that can represent owned, mutable platform-native strings, but is cheaply inter-convertible with Rust strings.

VaList	Experimental
A wrapper for a va_list

Enums

c_void
Equivalent to C's void type when used as a pointer.

c_void

Type Definitions

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

int8_t

int16_t

int32_t

int64_t

int128_t

intmax_t

intptr_t

ptrdiff_t

size_t

ssize_t

uint8_t

uint16_t

uint32_t

uint64_t

uint128_t

uintmax_t

uintptr_t